Making the module both TinyGo and Go friendly for compilation/testing. Tweaking registration functions.
