+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/wapc/wapc-guest-tinygo
+
+go 1.15
