+++ wapc.go
@@ -5,42 +5,6 @@ import (
 	"unsafe"
 )
 
-//go:wasm-module wapc
-//go:export __guest_request
-func guestRequest(operationPtr uintptr, payloadPtr uintptr)
-
-//go:wasm-module wapc
-//go:export __guest_response
-func guestResponse(ptr uintptr, len uint32)
-
-//go:wasm-module wapc
-//go:export __guest_error
-func guestError(ptr uintptr, len uint32)
-
-//go:wasm-module wapc
-//go:export __host_call
-func hostCall(
-	bindingPtr uintptr, bindingLen uint32,
-	namespacePtr uintptr, namespaceLen uint32,
-	operationPtr uintptr, operationLen uint32,
-	payloadPtr uintptr, payloadLen uint32) bool
-
-//go:wasm-module wapc
-//go:export __host_response_len
-func hostResponseLen() uint32
-
-//go:wasm-module wapc
-//go:export __host_response
-func hostResponse(ptr uintptr)
-
-//go:wasm-module wapc
-//go:export __host_error_len
-func hostErrorLen() uint32
-
-//go:wasm-module wapc
-//go:export __host_error
-func hostError(ptr uintptr)
-
 type (
 	// Function is the function to register in your waPC module.
 	Function func(payload []byte) ([]byte, error)
@@ -58,12 +22,18 @@ var (
 	allFunctions = Functions{}
 )
 
-// Register adds functions by name to the registery.
+// RegisterFunctions adds functions by name to the registery.
 // This should be invoked in `main()`.
-func Register(functions Functions) {
+func RegisterFunctions(functions Functions) {
 	for name, fn := range functions {
 		allFunctions[name] = fn
 	}
+}
+
+// RegisterFunction adds a single function by name to the registery.
+// This should be invoked in `main()`.
+func RegisterFunction(name string, fn Function) {
+	allFunctions[name] = fn
 }
 
 //go:export __guest_call
